time-voice-notion-deepseek/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dateutil==2.9.0
pydantic==2.7.1
requests==2.31.0
PyYAML==6.0.2
